Add support for Circuit Playground Express.
This change specifically adds support for the Adafruit Circuit
Playground Express. I bet it will allow other boards to be supported as
well. It uses different types when using the normal Arduino environment.
The ESP8266 environment does not define pin_size_t or PinStatus types
that are used in the normal Arduino environment. It might be useful to
try to get the ESP8266 Arduino environment to these types so that these
specializations aren't needed.